refactor(runtime): enforce subpath import contract (app/prompt/renderer)
