+++ apps/web/content/docs/core-concepts/commands.mdx
@@ -80,7 +80,7 @@ handler: async ({ flags, positional, shell, env, cwd, prompt, spinner, colors })
   // shell - Bun Shell for running commands (Bun.$)
   // env - Environment variables (process.env)
   // cwd - Current working directory
-  // prompt - Interactive prompts (auto-imported from @bunli/runtime)
+  // prompt - Interactive prompts (auto-imported from @bunli/runtime/prompt)
   // spinner - Progress indicators (auto-imported from @bunli/utils)
   // colors - Terminal colors (auto-imported from @bunli/utils)
 }
