+++ packages/runtime/src/index.ts
@@ -1,114 +1,13 @@
-// Canonical runtime surface. Prefer importing runtime and context APIs from @bunli/runtime.
+// Minimal root surface by design. Import value APIs from explicit subpaths.
 export type { OpenTuiRendererOptions, TuiRenderOptions } from './options.js'
-
-export {
-  RuntimeProvider,
-  useRuntime,
-  type RuntimeContextValue,
-  type RuntimeProviderProps,
-  RouteStoreProvider,
-  useRouteStore,
-  createInitialRouteState,
-  applyNavigate,
-  applyReplace,
-  applyBack,
-  canApplyBack,
-  type RouteStore,
-  type RouteState,
-  type RouteStoreProviderProps,
-  CommandRegistryProvider,
-  useCommandRegistry,
-  useCommandRegistryItems,
-  normalizeBinding,
-  commandToPaletteItem,
-  shouldCleanupRegisteredCommand,
-  type CommandRegistry,
-  type CommandRegistryProviderProps,
-  type RuntimeCommand,
-  type CommandPaletteItem
-} from './runtime/index.js'
-
-export {
-  ThemeProvider,
-  createTheme,
-  useTuiTheme,
-  darkThemeTokens,
-  lightThemeTokens,
-  type ThemeProviderProps,
-  type TuiTheme,
-  type TuiThemeInput,
-  type TuiThemeTokens,
-  createKeyMatcher,
-  matchesKeyBinding,
-  eventToBinding,
-  type KeyBinding,
-  type KeymapDefinition,
-  type KeyMatcher,
-  displayWidth,
-  truncateEnd,
-  padEndTo,
-  formatFixedWidth,
-  type TextOverflowMode,
-  FocusScopeProvider,
-  useFocusScope,
-  useScopedKeyboard,
-  dispatchScopedKeyboardEvent,
-  type FocusScopeProviderProps,
-  type ScopedKeyHandler,
-  type UseFocusScopeOptions,
-  type UseScopedKeyboardOptions,
-  OverlayHostProvider,
-  OverlayPortal,
-  type OverlayHostProviderProps,
-  type OverlayPortalProps,
-  DialogProvider,
-  useDialogManager,
-  DialogDismissedError,
-  sortDialogs,
-  getTopDialog,
-  getSelectableIndices,
-  getResolvedChooseIndex,
-  getAdjacentSelectableIndex,
-  type ChooseDialogOption,
-  type ChooseDialogOptions,
-  type ConfirmDialogOptions,
-  type DialogManager,
-  type DialogOpenOptions,
-  type DialogProviderProps,
-  type DialogRenderContext
-} from './components/index.js'
-
-export {
-  createPromptSession,
-  PromptCancelledError,
-  assertNotCancelled,
-  promptOrExit,
-  isCancel,
-  CANCEL,
-  text,
-  password,
-  confirm,
-  select,
-  multiselect,
-  group,
-  intro,
-  outro,
-  note,
-  log,
-  cancel,
-  rawSpinner,
-  type Cancel,
-  type PromptMode,
-  type PromptOptions,
-  type ConfirmOptions,
-  type SelectOption,
-  type SelectOptions,
-  type MultiSelectOptions,
-  type SpinnerAnimation,
-  type SpinnerOptions,
-  type Spinner,
-  type PromptApi,
-  type PromptSession,
-  type PromptSpinnerFactory,
-  type BunliPrompt
-} from './prompt/index.js'
+export type {
+  RuntimeRendererStartedEvent,
+  RuntimeRendererMissingRenderEvent,
+  RuntimeRendererDestroyedEvent,
+  RuntimePromptStartedEvent,
+  RuntimePromptCancelledEvent,
+  RuntimeTransportErrorEvent,
+  RuntimeEvent,
+  RuntimeEventType
+} from './events.js'
+export type { RuntimeTransport, RuntimeTransportObserver } from './transport.js'
