+++ packages/tui/src/components/command-palette.tsx
@@ -1,8 +1,8 @@
 import { useEffect, useId, useMemo, useState } from 'react'
-import { useScopedKeyboard } from '@bunli/runtime'
-import { createKeyMatcher } from '@bunli/runtime'
-import { useTuiTheme } from '@bunli/runtime'
-import { displayWidth, formatFixedWidth, type TextOverflowMode } from '@bunli/runtime'
+import { useScopedKeyboard } from '@bunli/runtime/app'
+import { createKeyMatcher } from '@bunli/runtime/app'
+import { useTuiTheme } from '@bunli/runtime/app'
+import { displayWidth, formatFixedWidth, type TextOverflowMode } from '@bunli/runtime/app'
 
 export interface CommandPaletteItem {
   key: string
