+++ packages/tui/src/components/empty-state.tsx
@@ -1,5 +1,5 @@
 import type { ReactNode } from 'react'
-import { useTuiTheme } from '@bunli/runtime'
+import { useTuiTheme } from '@bunli/runtime/app'
 
 export interface EmptyStateProps {
   title: string
