+++ packages/tui/src/components/modal.tsx
@@ -1,10 +1,10 @@
 import { useEffect, useId, useRef, type ReactNode } from 'react'
 import type { BoxRenderable, Renderable } from '@opentui/core'
 import { useRenderer } from '@opentui/react'
-import { createKeyMatcher } from '@bunli/runtime'
-import { useScopedKeyboard } from '@bunli/runtime'
-import { OverlayPortal } from '@bunli/runtime'
-import { useTuiTheme } from '@bunli/runtime'
+import { createKeyMatcher } from '@bunli/runtime/app'
+import { useScopedKeyboard } from '@bunli/runtime/app'
+import { OverlayPortal } from '@bunli/runtime/app'
+import { useTuiTheme } from '@bunli/runtime/app'
 
 export interface ModalProps {
   isOpen: boolean
