+++ packages/tui/src/components/multi-select-field.tsx
@@ -1,9 +1,9 @@
 import { useId, useMemo, useState } from 'react'
 import type { SelectOption } from '@opentui/core'
-import { useScopedKeyboard } from '@bunli/runtime'
+import { useScopedKeyboard } from '@bunli/runtime/app'
 import { useFormField } from './form-context.js'
-import { createKeyMatcher } from '@bunli/runtime'
-import { useTuiTheme } from '@bunli/runtime'
+import { createKeyMatcher } from '@bunli/runtime/app'
+import { useTuiTheme } from '@bunli/runtime/app'
 
 export interface MultiSelectOption extends SelectOption {
   label?: string
