+++ packages/tui/src/components/panel.tsx
@@ -1,5 +1,5 @@
 import type { ReactNode } from 'react'
-import { useTuiTheme } from '@bunli/runtime'
+import { useTuiTheme } from '@bunli/runtime/app'
 import {
   resolveVariantStyle,
   type ComponentEmphasis,
