+++ packages/tui/src/components/section-header.tsx
@@ -1,5 +1,5 @@
 import type { ReactNode } from 'react'
-import { useTuiTheme } from '@bunli/runtime'
+import { useTuiTheme } from '@bunli/runtime/app'
 
 export interface SectionHeaderProps {
   title: string
