+++ packages/tui/src/components/tabs.tsx
@@ -1,8 +1,8 @@
 import { useId, useMemo, useState } from 'react'
 import type { ReactNode } from 'react'
-import { useScopedKeyboard } from '@bunli/runtime'
-import { createKeyMatcher } from '@bunli/runtime'
-import { useTuiTheme } from '@bunli/runtime'
+import { useScopedKeyboard } from '@bunli/runtime/app'
+import { createKeyMatcher } from '@bunli/runtime/app'
+import { useTuiTheme } from '@bunli/runtime/app'
 
 export interface TabItem {
   key: string
