+++ skills/bunli/SKILL.md
@@ -256,7 +256,7 @@ handler: async ({ prompt }) => {
 |---------|---------|
 | `@bunli/core` | CLI framework (defineCommand, option, createCLI) |
 | `@bunli/utils` | Colors and validation utilities |
-| `@bunli/runtime` | Prompt and spinner APIs |
+| `@bunli/runtime/prompt` | Prompt and spinner APIs |
 | `@bunli/tui` | Terminal UI components |
 | `bunli` | CLI for building CLIs |
 | `create-bunli` | Project scaffolding |
@@ -277,7 +277,7 @@ Bunli uses **OpenTUI** as its terminal rendering engine. Understanding when to u
 - Building CLI applications with commands and options
 - Plugin architecture (auth, config, completions)
 - Type-safe CLI with Zod validation
-- Interactive prompts via `@bunli/runtime`
+- Interactive prompts via `@bunli/runtime/prompt`
 - Bunli TUI components (`Form`, `SchemaForm`, `DataTable`, `ProgressBar`, and more)
 - Publishing CLI to npm
 
@@ -305,7 +305,7 @@ import { createCliRenderer } from "@opentui/core"
 **Package relationship:**
 - `@bunli/tui` wraps `@opentui/react`
 - Bunli auto-wires the OpenTUI renderer runtime for `render` commands
-- `@bunli/runtime` provides prompt + spinner APIs used in handlers
+- `@bunli/runtime/prompt` provides prompt + spinner APIs used in handlers
 - Bunli TUI hooks (`useKeyboard`, `useTimeline`) are re-exported from OpenTUI
 - Renderer options map to OpenTUI renderer settings
